Combine top-level flask setup in app.py
